+++ src/components/AdvancedInnerTable/NumberCell/NumberCell.tsx
@@ -1,8 +1,8 @@
 import { styled } from '@mui/material';
 import React from 'react';
 
-import { formatNumber } from '../../core/utils/string';
-import type { RowType } from '../AdvancedInnerTable/types';
+import { formatNumber } from '../../../core/utils/string';
+import type { RowType } from '../types';
 
 interface NumberCellProps {
   value: number;
