+++ src/components/AdvancedInnerTable/TextCell/TextCell.tsx
