+++ src/views/CoreUnitBudgetStatement/transparencyReportUtils.tsx
@@ -9,13 +9,13 @@ import { getPageWrapper } from '@ses/core/utils/dom';
 import lightTheme from '@ses/styles/theme/themes';
 import MobileDetect from 'mobile-detect';
 import React, { useCallback, useEffect, useRef, useState } from 'react';
-import { NumberCell } from '@/components/NumberCell/NumberCell';
+import { NumberCell } from '@/components/AdvancedInnerTable/NumberCell/NumberCell';
+import { TextCell } from '@/components/AdvancedInnerTable/TextCell/TextCell';
 import { WalletTableCell } from '@/components/WalletTableCell/WalletTableCell';
 import Information from '@/components/icons/information';
 import ArrowPopoverTargetValueComponent from '@/views/CoreUnitBudgetStatement/components/ArrowPopoverTargetValue/ArrowPopoverTargetValueComponent';
 import { formatAddressForOutput } from '../../core/utils/string';
 import { CustomLink } from '../../stories/components/CustomLink/CustomLink';
-import { TextCell } from '../../stories/components/TextCell/TextCell';
 import ModalSheetValueContent from './components/TransparencyTransferRequest/components/ModalSheet/ModalSheetValueContent';
 import type { BudgetStatementWalletDto } from '../../core/models/dto/coreUnitDTO';
 import type { TargetBalanceTooltipInformation } from '@ses/core/utils/typesHelpers';
